20191012 빌드 테스트 2015버전 다운 & 경고문구 수정

+++ Chat/Server/Server.cpp
@@ -207,7 +207,7 @@ unsigned WINAPI RecvThread(LPVOID pCompPort) {
 
 			{
 				lock_guard<mutex> guard(queueCs);
-				packetCnt.fetch_add(packetQueue.size());
+				packetCnt.fetch_add(static_cast<int>(packetQueue.size()));
 				while (!packetQueue.empty()) { // packetQueue -> jobQueue 
 					JOB_DATA jobData = packetQueue.front();
 					packetQueue.pop();
